requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
defusedxml==0.7.1
freezegun==1.2.1
idna==3.3
importlib-metadata==4.11.3
jira==3.2.0
keyring==23.5.0
oauthlib==3.2.0
psycopg2==2.9.3
pytest==7.1.2
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2019.3
requests==2.27.1
requests-mock==1.9.3
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
setuptools==54.1.2
six==1.16.0
typing_extensions==4.2.0
urllib3==1.26.9
zipp==3.8.0
